+++ CHANGES.md
@@ -3,6 +3,7 @@
 # 0.3.1
 
 * Get artifacts of a build
+* Support external jobs
 
 # 0.3.0 (2018/05/13)